uint32 updates

+++ include/index.hpp
@@ -29,11 +29,12 @@ struct index {
     uint64_t u2c(uint64_t unitig_id) const { return m_u2c.rank(unitig_id); }
 
     void pseudoalign_full_intersection(std::string const& sequence,
-                                       std::vector<uint32_t>& results) const;
-    void pseudoalign_threshold_union(std::string const& sequence, std::vector<uint32_t>& results,
+                                       std::vector<uint64_t>& results) const;
+
+    void pseudoalign_threshold_union(std::string const& sequence, std::vector<uint64_t>& results,
                                      const double threshold) const;
 
-    void intersect_unitigs(std::vector<uint32_t>& unitig_ids, std::vector<uint32_t>& colors) const;
+    void intersect_unitigs(std::vector<uint64_t>& unitig_ids, std::vector<uint32_t>& colors) const;
 
     uint64_t num_bits() const {
         return m_k2u.num_bits() + m_u2c.bytes() * 8 + m_ccs.num_bits() + m_filenames.num_bits();
